+++ go.mod
@@ -1,16 +1,16 @@
-module proj
+module iot-device-tracker-service
 
 go 1.20
 
 require (
 	github.com/joho/godotenv v1.4.0
+	github.com/rs/zerolog v1.26.1
 	google.golang.org/grpc v1.48.0
 	google.golang.org/protobuf v1.28.1
 )
 
 require (
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/rs/zerolog v1.26.1 // indirect
 	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
 	golang.org/x/sys v0.0.0-20210809222454-d867a43fc93e // indirect
 	golang.org/x/text v0.3.6 // indirect
